Trabalhando em diversas issues relacionadas:
* fixed #51
* fixed #52
* fixed #53

+++ pedidos/models.py
@@ -33,7 +33,7 @@ class Pedidos(models.Model):
 
     datapedido = models.DateField(db_column="datapedido", blank=True, null=True)
     valorpedido = models.DecimalField(
-        db_column="valorpedido", blank=True, null=True, max_digits=6, decimal_places=2
+        db_column="valorpedido", blank=True, null=True, max_digits=10, decimal_places=2
     )
     observacao = models.CharField(
         db_column="observacao", max_length=255, blank=True, null=True
@@ -81,7 +81,9 @@ class Produtos(models.Model):
     descricao = models.CharField(
         db_column="descricao", max_length=255, blank=True, null=True
     )
-    valorcusto = models.FloatField(db_column="valorcusto", blank=True, null=True)
+    valorcusto = models.DecimalField(
+        db_column="valorcusto", blank=True, null=True, max_digits=10, decimal_places=2
+    )
 
     class Meta:
         managed = False
